get params from BBDD and use params + type jobs result

+++ src/services/JobsManagerService.ts
@@ -1,8 +1,11 @@
 const schedule = require('node-schedule');
+import settings from '../../settings.json'
 import { CollectionsHelper } from '../helpers/collectionsHelper';
 import { LogService } from './LogService/LogService';
 import { LogType } from './LogService/LogTypeEnum';
 import { JobsModel } from './Jobs.model';
+import { GetAllCollectionsJobResult } from './GetAllCollectionsJobResult';
+import { GetAllCollectionsStatsJobResult } from './GetAllCollectionsStatsJobResult';
 
 export class JobsManagerService {
     private _collectionsHelper: CollectionsHelper;
@@ -30,8 +33,14 @@ export class JobsManagerService {
     }
 
     public startJobs = async() => {
-        await this._getAllCollectionsJob();
-        await this._getAllCollectionsStatsDataJob();
+        if (settings.JobsManager.GetAllCollectionsJobEnable) {
+            await this._getAllCollectionsJob();
+        }
+
+        if (settings.JobsManager.GetAllCollectionsStatsDataJobEnable) {
+            await this._getAllCollectionsStatsDataJob();
+        }
+
         this._startScheduleJobs();
     }
 
@@ -44,8 +53,10 @@ export class JobsManagerService {
     private _startDeleteAllTracesJob = (): void => {
         try {
             // Cada 96 horas (4 días)
-            const job: any = schedule.scheduleJob('0 */96 * * *', () => {
-                this._deleteAllTracesJob();
+            const job: any = schedule.scheduleJob(settings.JobsManager.DeleteAllTracesJobTime, () => {
+                if (settings.JobsManager.DeleteAllTracesJobEnable) {
+                    this._deleteAllTracesJob();
+                }
             });
         } catch (error) {
         }
@@ -61,12 +72,14 @@ export class JobsManagerService {
     private _startGetAllCollectionsJob = (): void => {
         try {
             // Cada 3 horas
-            const job: any = schedule.scheduleJob('0 */3 * * *', () => {
-                if (this._allCollectionsStatsDataJob.Working) {
-                    this._allCollectionsJob.Pending = true;
-                } else {
-                    this._allCollectionsJob.Working = true;
-                    this._getAllCollectionsJob();
+            const job: any = schedule.scheduleJob(settings.JobsManager.GetAllCollectionsJobTime, () => {
+                if (settings.JobsManager.GetAllCollectionsJobEnable) {
+                    if (this._allCollectionsStatsDataJob.Working) {
+                        this._allCollectionsJob.Pending = true;
+                    } else {
+                        this._allCollectionsJob.Working = true;
+                        this._getAllCollectionsJob();
+                    }
                 }
             });
         } catch (error) {
@@ -78,17 +91,17 @@ export class JobsManagerService {
         const startTime: number = Date.now() / 1000;
         this._logService.log(this._allCollectionsJob.StartLog, LogType.Information);
         return new Promise((resolve, reject) => {
-            this._collectionsHelper.getAllCollections().then(result => {
+            this._collectionsHelper.getAllCollections().then((result: GetAllCollectionsJobResult) => {
                 this._manageGetAllCollectionsJobResult(startTime, result);
                 resolve(true);
-            }).catch(error => {
+            }).catch((error: string) => {
                 this._manageGetAllCollectionsJobError(error);
                 reject(false)
             });
         })
     }
 
-    private _manageGetAllCollectionsJobResult = (startTime: number, result: any): void => {
+    private _manageGetAllCollectionsJobResult = (startTime: number, result: GetAllCollectionsJobResult): void => {
         this._allCollectionsJob.Working = false;
         const endTime: string = (Date.now() / 1000 - startTime).toFixed(2) + " segundos";
         const dataTrace: string = "TIME: " + endTime + ". NewCollectionsToSave: " + result?.NewCollectionsToSave + ". NewCollectionsSaved: " + result?.NewCollectionsSaved + ". AllCollections: " + result?.AllCollections;
@@ -102,7 +115,7 @@ export class JobsManagerService {
         this._manageGetAllCollectionsStatsDataPendingJobs();
     }
 
-    private _manageGetAllCollectionsJobError = (error: any): void => {
+    private _manageGetAllCollectionsJobError = (error: string): void => {
         this._allCollectionsJob.Working = false;
         this._logService.log(this._allCollectionsJob.EndLog + "With ERRORS", LogType.Error);
         this._logService.log(this._allCollectionsJob.ErrorLog + error, LogType.Error);
@@ -120,12 +133,14 @@ export class JobsManagerService {
     private _startGetAllCollectionsStatsDataJob = (): void => {
         try {
             // En el minuto 10 cada 8 horas
-            const job: any = schedule.scheduleJob('15 */8 * * *', () => {
-                if (this._allCollectionsJob.Working) {
-                    this._allCollectionsStatsDataJob.Pending = true;
-                } else {
-                    this._allCollectionsStatsDataJob.Working = true;
-                    this._getAllCollectionsStatsDataJob();
+            const job: any = schedule.scheduleJob(settings.JobsManager.GetAllCollectionsStatsDataJobTime, () => {
+                if (settings.JobsManager.GetAllCollectionsStatsDataJobEnable) {
+                    if (this._allCollectionsJob.Working) {
+                        this._allCollectionsStatsDataJob.Pending = true;
+                    } else {
+                        this._allCollectionsStatsDataJob.Working = true;
+                        this._getAllCollectionsStatsDataJob();
+                    }
                 }
             });
         }
@@ -138,17 +153,17 @@ export class JobsManagerService {
         const startTime: number = Date.now() / 1000;
         this._logService.log(this._allCollectionsStatsDataJob.StartLog, LogType.Information);
         return new Promise((resolve, reject) => {
-            this._collectionsHelper.getAllCollectionsStatsData().then(result => {
+            this._collectionsHelper.getAllCollectionsStatsData().then((result: GetAllCollectionsStatsJobResult) => {
                 this._manageGetAllCollectionsStatsDataJobResult(startTime, result);
                 resolve(true);
-            }).catch(error => {
+            }).catch((error: string) => {
                 this._manageGetAllCollectionsStatsDataJobError(error);
                 reject(false)
             });
         })
     }
 
-    private _manageGetAllCollectionsStatsDataJobResult = (startTime: number, result: any): void => {
+    private _manageGetAllCollectionsStatsDataJobResult = (startTime: number, result: GetAllCollectionsStatsJobResult): void => {
         this._allCollectionsStatsDataJob.Working = false;
         const endTime: string = (Date.now() / 1000 - startTime).toFixed(2) + " segundos";
         const dataTrace: string = "TIME: " + endTime + ". ActiveAllCollections: " + result?.ActiveAllCollections + ". CollectionsUpdated: " + result?.CollectionsUpdated;
@@ -162,7 +177,7 @@ export class JobsManagerService {
         this._manageGetAllCollectionsPendingJobs();
     }
 
-    private _manageGetAllCollectionsStatsDataJobError = (error: any): void => {
+    private _manageGetAllCollectionsStatsDataJobError = (error: string): void => {
         this._allCollectionsStatsDataJob.Working = false;
         this._logService.log(this._allCollectionsStatsDataJob.EndLog + "With ERRORS", LogType.Error);
         this._logService.log(this._allCollectionsStatsDataJob.ErrorLog + error, LogType.Error);
